ci: Get coverity running again accounting for changes after the outtage.
It looks like there have been some upstream changes in the way coverity
and travis are integrated. Previously the coverity scan build was done
after all other build / test stuff. This made sense since we only want
to fix issues found by coverity for builds that actually build. After
the scan.coverity.com service came back online, the coverity builds
happen before all other tasks in the CI.

Given all of this we have to change a few things around. This commit
adds the necessary commands to bootstrap the source tree, then do a
build instrumented with the coverity stuff in a VPATH. Doing the
coverity build in a VPATH is necessary as it makes the other builds
done by the CI *after* the scan build possible. Autotools doesn't like
configuring a VPATH build from a source tree that has already been
configured.

All scan tokens appear to have been refreshed and so we've created a new
encrypted variable for it. Finally the scan service is also rejecting
access if the `name` for the project is reported as anything other than
`01org/TPM2.0-TSS`.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>